+++ examples/averageBehaviour.c
@@ -17,7 +17,7 @@
 */
 
 #include <stdio.h>
-#include "../include/cgp.h"
+#include "../src/cgp.h"
 
 int main(void){
 
@@ -34,7 +34,7 @@ int main(void){
 	int numGens = 10000;
 	int numRuns = 10;
 	
-	float targetFitness = 0.1;
+	double targetFitness = 0.1;
 	int updateFrequency = 500;
 
 	int averageFitness;
